requirements.txt
./wheels/google_alerts-0.3-py3-none-any.whl
alabaster==0.7.10
annotated-types==0.7.0
Babel==2.5.3
beautifulsoup4==4.12.3
bs4==0.0.1
cachetools==5.3.3
certifi==2018.4.16
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
docutils==0.14
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.123.0
google-auth==2.30.0
google-auth-httplib2==0.1.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.18.1
idna==2.6
imagesize==1.0.0
inflection==0.5.1
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==17.1
pkginfo==1.4.2
proto-plus==1.24.0
protobuf==4.25.3
pyairtable==2.3.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.2.0
pyparsing==2.2.0
pytz>=2018.4
requests==2.32.3
requests-toolbelt==0.8.0
rsa==4.9
selenium==3.12.0
setuptools==70.1.1
six==1.16.0
snowballstemmer==1.2.1
soupsieve==2.5
Sphinx==1.7.4
sphinx-rtd-theme==0.3.0
sphinxcontrib-websupport==1.0.1
twine==1.11.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
feedparser==6.0.11
python-dateutil==2.9.0
markdownify==0.12.1
pandas==2.2.2
pickle-function-cache==0.1.5
halo==0.0.31
playwright==1.44.0

